sprint-3/step-1 feat: вынести в MainPage начинку основной страницы
