--- src/components/App/App.module.css
@@ -1,21 +1,0 @@
-.application {
-  width: 100%;
-  color: hsla(240, 5%, 95%, 1);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-}
-
-.main {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  gap: 40px;
-}
-
-.errorText {
-  color: hsla(5, 80%, 50%, 1);
-}
